pyproject.toml
[tool.poetry]
name = "fmsolver"
version = "0.1.0"
description = "Simple brute-force team solver"
authors = ["James Robinson <james.em.robinson@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
